1/requirements.txt
torch>=1.8.0
torchvision>=0.9.0
pillow>=8.0.0
scikit-learn>=0.24.0
pandas>=1.1.0
joblib>=1.0.0
tqdm>=4.0.0
